+++ pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.cs</groupId>
-	<artifactId>enodes</artifactId>
+	<artifactId>enotes</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>enodes</name>
+	<name>enotes</name>
 	<description>Demo project for Spring Boot</description>
 	<url/>
 	<licenses>
@@ -27,7 +27,7 @@
 		<url/>
 	</scm>
 	<properties>
-		<java.version>21</java.version>
+		<java.version>23</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
