+++ src/main/java/com/cs/enotes/entity/BaseModel.java
@@ -1,0 +1,21 @@
+package com.cs.enotes.entity;
+
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@MappedSuperclass
+public class BaseModel {
+
+    private String description;
+    private Boolean isActive;
+    private Boolean isDeleted;
+    private Integer createdBy;
+    private Date createdOn;
+    private Integer updatedBy;
+    private Date updatedOn;
+}
