+++ src/main/resources/application.yml
@@ -1,0 +1,23 @@
+server:
+  port: 8080
+#  servlet:
+#    session:
+#      persistent: false
+spring:
+  application:
+    name: e-notes
+#   threads:
+#     virtual:
+#       enabled: true
+  datasource:
+    url: jdbc:mysql://localhost:3306/testing-pro?createDatabaseIfNotExist=true
+    username: root
+    password: root
+  jpa:
+    hibernate:
+      naming:
+        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+      ddl-auto: none
+  sql:
+    init:
+      mode: always
