+++ src/test/java/com/cs/enotes/EnodesApplicationTests.java
@@ -1,4 +1,4 @@
-package com.cs.enodes;
+package com.cs.enotes;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
